+++ NEWS.md
@@ -1,5 +1,7 @@
 # textures 0.0.0.9005
 
+* New function `png_plot3d()` to directly plot from a PNG file. 
+
 * Modify R functions to leverage C++ versions of quad vertex and quad index constructors. 
 
 * Aligned to R's matrix orientation as a basis for quad creation.